build: refactor and complete unit test
