--- statemigration/test_config.go
@@ -1,8 +1,0 @@
-package statemigration
-
-import "testing"
-
-// TODO: Implement this unit test
-func TestGroupToVariablesDecode(t *testing.T) {
-	
-}
